database.py: Database option added
Instead of using queues for publishes, a database can be used.
The line

  "database": true

must be added to the connect.cfg file and

    client.config.database_file = database_file

must be added to the app to use the database.
Otherwise, the original queue method will be used.
If the database storage limit is set and reached,
it will default to the queue method.

Signed-off-by: lcc755 <lcc755@mun.ca>