client/handler: add support to read current telemetry/attribute
This commit adds support for a new set of APIs that read back the
current telemetry or attribute sample from the cloud.  Formerly, the
only way to do this was via REST API, which was not an option for the
device since it required login credentials.

The new APIs are:
  * client.telemetry_read_last_sample(name)
  * client.attribute_read_last_sample(name)

Note: if a tight loop of publish and read is used, make sure
to poll for the status of the read result. If it is non 0, read again.
See iot-sample-telemetry.py for details.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>