handler.py: add network check loop to connect
Resolves #169
Previously, if there was no network when starting the device manager,
the device manager would exit.  Daemons like the device manager are
typically handled by a monitoring service such as (e.g.) systemd.
However, not every system has a monitor.  To help with this scenario,
a network test was added to the connect API which essentially loops
until gethostbyname succeeds.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>